Add base for MultiSelect fields and organize fields package
