+++ swimlane/core/fields/base/field.py
@@ -1,0 +1,91 @@
+import weakref
+
+
+class Field(object):
+    """Base class for abstracting Swimlane complex types"""
+
+    field_type = None
+
+    # Sentinel representing a field that has no current value
+    _unset = object()
+
+    # List of supported types, leave blank to disable type validation
+    supported_types = []
+
+    def __init__(self, name, record, default_value=_unset, allow_null=True):
+        self.name = name
+        self.__record_ref = weakref.ref(record)
+        self.default = default_value
+        self.allow_null = allow_null
+        self._value = self.default
+
+        self.field_definition = self.record._app.get_field_definition_by_name(self.name)
+        self.id = self.field_definition['id']
+        self.input_type = self.field_definition.get('inputType')
+
+    def __repr__(self):
+        return '<{class_name}: {py!r}>'.format(class_name=self.__class__.__name__, py=self.get_python())
+
+    @property
+    def record(self):
+        return self.__record_ref()
+
+    def _get(self):
+        """Default getter used for both representations unless overridden"""
+        return self._value
+
+    def get_python(self):
+        """Return best python representation of field value"""
+        return self._get()
+
+    def cast_to_swimlane(self, value):
+        """Called during get_swimlane, should accept a python value and return swimlane representation
+        
+        Defaults to no-op
+        """
+        return value
+
+    def get_swimlane(self):
+        """Return best swimlane representation of field value"""
+        return self.cast_to_swimlane(self._get())
+
+    def validate_value(self, value):
+        """Validate value is an acceptable type during _set operation"""
+        if value is None:
+            if not self.allow_null:
+                raise ValueError('Field "{}" does not allow null and cannot be set to None')
+        else:
+            if self.supported_types and not isinstance(value, tuple(self.supported_types)):
+                raise TypeError('Field "{}" expects one of "{}", got "{}" instead'.format(
+                    self.name,
+                    ', '.join([t.__name__ for t in self.supported_types]),
+                    type(value).__name__)
+                )
+
+    def _set(self, value):
+        """Default setter used for both representations unless overridden"""
+        self.validate_value(value)
+        self._value = value
+
+    def set_python(self, value):
+        """Set field internal value from the python representation of field value"""
+        return_value = self._set(value)
+
+        self.record._raw['values'][self.id] = self.get_swimlane()
+
+        return return_value
+
+    def cast_to_python(self, value):
+        """Called during set_swimlane, should accept a single raw value as provided from API
+        
+        Defaults to no-op
+        """
+        return value
+
+    def set_swimlane(self, value):
+        """Set field internal value from the swimlane representation of field value"""
+        return self._set(self.cast_to_python(value))
+
+    def unset(self):
+        """Marks field as having no particular value"""
+        self._value = self.default
