+++ swimlane/core/fields/tracking.py
@@ -1,0 +1,6 @@
+from .base import ReadOnly, Field
+
+
+class TrackingField(ReadOnly, Field):
+
+    field_type = 'Core.Models.Fields.TrackingField, Core'
